pyproject.toml
[tool.poetry]
name = "gcf"
version = "0.1.0"
description = ""
authors = ["Janne Enberg <janne.enberg@lietu.net>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-firestore = "2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
